+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+### Changed
+
+- Fixed allowed `-webkit-` values and properties
+
 ## [8.0.0] - 2024-02-21
 
 ### Added